requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-docx>=1.1.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
reportlab>=4.0.0